refactor: group shrinkage models and estimators
