+++ cicd/tests/test_phantom_eval_jax.py
@@ -4,8 +4,8 @@ import jax
 import numpy as np
 from jax import numpy as jnp
 
-import jaxns.phantom_eval as jax_phantom
-import jaxns.phantom_eval_ref as ref_phantom
+import jaxns.shrinkage.phantom as jax_phantom
+import jaxns.shrinkage.reference as ref_phantom
 from jaxns.algorithm.race_tree import BlockState
