+++ src/jaxns/shrinkage/reference.py
